android: implement getStaticByteField

+++ unidbg-android/src/main/java/com/github/unidbg/linux/file/ByteArrayFileIO.java
@@ -38,10 +38,10 @@ public class ByteArrayFileIO extends BaseAndroidFileIO {
     }
 
     @Override
-    public int pread(Backend backend, Pointer buffer, int count, int offset) {
+    public int pread(Backend backend, Pointer buffer, int count, long offset) {
         int pos = this.pos;
         try {
-            this.pos = offset;
+            this.pos = (int) offset;
             return read(backend, buffer, count);
         } finally {
             this.pos = pos;
