+++ unidbg-api/src/main/java/com/github/unidbg/file/FileIO.java
@@ -18,7 +18,7 @@ public interface FileIO {
 
     int read(Backend backend, Pointer buffer, int count);
 
-    int pread(Backend backend, Pointer buffer, int count, int offset);
+    int pread(Backend backend, Pointer buffer, int count, long offset);
 
     int fcntl(Emulator<?> emulator, int cmd, long arg);
